feat: create a welcome screen with "create account" and "sign in" buttons
This welcome screen should be the first screen the user ever sees when
opening Habitopia. For now, the logo and background image are placeholder
images. There are two buttons for creating an account and signing in which
do not work yet. There is also the Habitopia heading and subheading with
the tagline.